Abuse Prevention
Massive updates to the functions in music.py to prevent people not in the same voice channel as the bot from modifying the bot. this does not affect the play command. A more extensive implementation will be needed to cover play. I also did not do queue as I did not think it was neccessary. Can add in the future if it proves to be something needed.