Clearing of cache allowed, Welcome function fixes
This has allowed the use of a new function to clear the bot if it is misbehaving. This should not used often and only if the bot is consistently behaving poorly. The read me has more information on this. The welcome function now requires discord intents. this is reflected in the code and must also be changed in the discord developers page, more information in the read me.