Pause Upgrade
This now allows the pause command to wait seven minutes after being called to un-pause automatically. This works well but I believe it can have undesired effects if you pause, un-pause and the pause again. The first un-pause timer will cause the second pause to end before the 7 minute mark. This needs to be looked into but is not a critical issue.