Playlist format update
Viewplaylist and Listplaylist have had an update on how the contents are displayed. lpl is now broken into pages of 10 that can be selected by an index to prevent the error of having too many to where discord will no longer display them. Viewplaylist has a similar update however it will automatically show all of the songs in chunks of 20.