Shuffle and Playlist update
New custom shuffle function that shuffles in a "one time" fashion instead of a live shuffle. It is a custom function and not one present native like what was used prior. Also there is a new function by popular demand to add the entire queue to a given playlist name.